Conflicts:
	src/app/app.component.ts
	src/app/app.module.ts
	src/app/catalogue/product-publish.component.ts
	src/app/explorative-search/explorative-search-details.component.ts

+++ src/app/catalogue/ubl-model-view/catalogue/catalogue-line-panel.component.ts
@@ -2,6 +2,8 @@ import {Component, Input} from "@angular/core";
 import {CatalogueLine} from "../../model/publish/catalogue-line";
 import {CatalogueService} from "../../catalogue.service";
 import {Router} from "@angular/router";
+import {PublishService} from "../../publish-and-aip.service";
+import {CategoryService} from "../../category/category.service";
 
 @Component({
     selector: 'catalogue-line-panel',
@@ -14,15 +16,22 @@ export class CatalogueLinePanelComponent {
     @Input() presentationMode: string;
 
     constructor(private catalogueService: CatalogueService,
+                private categoryService: CategoryService,
+                private publishService: PublishService,
                 private router: Router) {
     }
 
     redirectToEdit() {
         this.catalogueService.editCatalogueLine(this.catalogueLine);
-        this.router.navigate(['publish'], {queryParams: {pageRef: "catalogue"}});
+        this.publishService.publishMode = 'edit';
+        this.publishService.publishingStarted = false;
+        this.categoryService.resetSelectedCategories();
+        this.router.navigate(['catalogue/publish']);
     }
 
     deleteCatalogueLine(): void {
-        this.catalogueService.deleteCatalogueLine(this.catalogueService.catalogue.uuid, this.catalogueLine.id);
+		if (confirm("Are you sure that you want to delete this catalogue item?")) {
+			this.catalogueService.deleteCatalogueLine(this.catalogueService.catalogue.uuid, this.catalogueLine.id);
+		}
     }
 }
